+++ public/board/3/index.html
@@ -6,7 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Tic-Tac-Toe Game Play</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="../../styles/board-3.css">
     <link type="image/png" sizes="50x50" rel="icon" href="https://img.icons8.com/ios/50/000000/play--v3.png">
 </head>
 
@@ -47,7 +47,7 @@
         </div>
     </div>
 
-    <script type="module" src="app.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 
 </html>
